Add custom hook for fetching the meals
